old_version/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "vgcn_bert"
version = "1.1.0"
description = "VGCN-BERT: Augmenting BERT with Graph Embedding"
requires-python = ">=3.9"
dependencies = [
  "pandas==1.4.*",
  "numpy==1.21.*",
  "pytorch>=1.0",
  "scikit-learn==1.1.*",
  "transformers==0.6.2",
  "nltk",
  "tqdm",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "isort==5.10.1",
  "black==22.3.0",
  "mypy==0.991",
  "pytest==7.*",
  "pytest-cov",
  "pre-commit==2.20.*",
  "jupyterlab==3.4.*",
]

[project.scripts]
vgcn_bert_prepare_data = "vgcn_bert.cli.prepare_data:cli_main"

[tool.setuptools.packages]
find = {}

[tool.black]
line-length = 79
target-version = ['py39']
include = '''
(
    ^/tests/.*\.pyi?$
  | ^/vgcn_bert/.*\.pyi?$
  | ^/snippets/.*\.pyi?$
  | ^/setup[.]py
  | ^/conftest[.]py
)
'''
exclude = '''
(  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | venv
  | __pycache__
)
'''

[tool.pycln]
include = '''
(
    ^/tests/
  | ^/vgcn_bert/
  | ^/snippets/
)
'''
all = true

[tool.isort]
profile = "black"
src_paths = ["vgcn_bert", "tests", "snippets" ]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79


[tool.mypy]
files = '''
(
    ^/vgcn_bert/.*\.pyi?$
    | ^/tests/.*\.pyi?$
    | ^/snippets/.*\.pyi?$
)
'''
python_version = 3.9
[[tool.mypy.overrides]]
module = 'pytest'
ignore_missing_imports = true
